# Speedup ParseError merging