Before

| Method | Mean           | Error        | StdDev       |
|------- |---------------:|-------------:|-------------:|
| Parse  | 4,223,378.5 us | 53,554.05 us | 47,474.27 us |
| Lex    |       819.4 us |     16.15 us |     13.48 us |

After

| Method | Mean           | Error        | StdDev       |
|------- |---------------:|-------------:|-------------:|
| Parse  | 3,840,863.4 us | 18,454.98 us | 15,410.74 us |
| Lex    |       758.0 us |      5.56 us |      4.64 us |